refactor eos into task
