+++ mlip_arena/tasks/eos.py
@@ -13,9 +13,9 @@ from ase import Atoms
 from ase.filters import *  # type: ignore
 from ase.optimize import *  # type: ignore
 from ase.optimize.optimize import Optimizer
-from prefect import flow
+from prefect import task
 from prefect.futures import wait
-from prefect.runtime import flow_run, task_run
+from prefect.runtime import task_run
 from pymatgen.analysis.eos import BirchMurnaghan
 
 from mlip_arena.models import MLIPEnum
@@ -25,20 +25,8 @@ if TYPE_CHECKING:
     from ase.filters import Filter
 
 
-def generate_flow_run_name():
-    flow_name = flow_run.flow_name
-
-    parameters = flow_run.parameters
-
-    atoms = parameters["atoms"]
-    calculator_name = parameters["calculator_name"]
-
-    return f"{flow_name}: {atoms.get_chemical_formula()} - {calculator_name}"
-
-
 def generate_task_run_name():
     task_name = task_run.task_name
-
     parameters = task_run.parameters
 
     atoms = parameters["atoms"]
@@ -47,10 +35,8 @@ def generate_task_run_name():
     return f"{task_name}: {atoms.get_chemical_formula()} - {calculator_name}"
 
 
-# https://docs.prefect.io/3.0/develop/write-tasks#custom-retry-behavior
-# @task(task_run_name=generate_task_run_name)
-@flow(flow_run_name=generate_flow_run_name, validate_parameters=False)
-def fit(
+@task(task_run_name=generate_task_run_name)
+def run(
     atoms: Atoms,
     calculator_name: str | MLIPEnum,
     calculator_kwargs: dict | None,
