+++ tests/test_eos.py
@@ -2,14 +2,38 @@ import sys
 
 import pytest
 from ase.build import bulk
+from prefect import flow
 from prefect.testing.utilities import prefect_test_harness
 
 from mlip_arena.models import MLIPEnum
-from mlip_arena.tasks.eos.run import fit as EOS
+from mlip_arena.tasks.eos import run as EOS
 
-atoms = bulk("Cu", "fcc", a=3.6)
 
-@pytest.mark.skipif(sys.version_info[:2] != (3,11), reason="avoid prefect race condition on concurrent tasks")
+@flow
+def single_eos_flow(calculator_name):
+    atoms = bulk("Cu", "fcc", a=3.6)
+
+    return EOS(
+        atoms=atoms,
+        calculator_name=calculator_name,
+        calculator_kwargs={},
+        device=None,
+        optimizer="BFGSLineSearch",
+        optimizer_kwargs=None,
+        filter="FrechetCell",
+        filter_kwargs=None,
+        criterion=dict(
+            fmax=0.1,
+        ),
+        max_abs_strain=0.1,
+        npoints=6,
+    )
+
+
+@pytest.mark.skipif(
+    sys.version_info[:2] != (3, 11),
+    reason="avoid prefect race condition on concurrent tasks",
+)
 @pytest.mark.parametrize("model", [MLIPEnum["MACE-MP(M)"]])
 def test_eos(model: MLIPEnum):
     """
@@ -17,21 +41,8 @@ def test_eos(model: MLIPEnum):
     """
 
     with prefect_test_harness():
-
-        result = EOS(
-            atoms=atoms,
+        result = single_eos_flow(
             calculator_name=model.name,
-            calculator_kwargs={},
-            device=None,
-            optimizer="BFGSLineSearch",
-            optimizer_kwargs=None,
-            filter="FrechetCell",
-            filter_kwargs=None,
-            criterion=dict(
-                fmax=0.1,
-            ),
-            max_abs_strain=0.1,
-            npoints=6,
         )
 
         assert isinstance(result["K"], float)
